extend integration tests
